requirements.txt
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.15.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
parso==0.8.4
psycopg2-binary==2.9.9
pycparser==2.22
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
WTForms==3.1.2
